Aircraft type as per ICAO.

A dataset containing aircraft types as "published" at ICAO's Aicraft Type Designator page.

aircrafttype

Format

A data frame with 9290 rows and 19 variables. The variables are a joint of flight information from `flts` with the following ones:

ModelFullName

Full name of the aircraft model

Description

Description of the aircraft type. The first character is L = landplane, S = seaplane, A = amphibian, H = helicopter, G = gyrocopter, T = tilt-wing aircraft. The second character is about the number of engines 1 thru 8 or C applicable to fixed-wing aircraft only and indicating that two engines are coupled to drive a single propeller system. The fird character is for the type of engine with P = piston engine, T = turboprop/turboshaft engine, J = jet engine, E = electric engine

WTC

The Wake Turbulence Category (WTC) indicator defined according to the certificated maximum take-off mass (MTOM). H = heavy (300000 lb), M = medium (less than 136000 kg (300000 lb) and more than 7000 kg (15500 lb))

Designator

The aircraft type designator

ManufacturerCode

The aircraft manufacturer code

AircraftDescription

The aircraft kind, see the first character of the Description field.

EngineCount

The number of engines, see the second character of the Description field.

EngineType

The engine type, see the third character of the Description field.

Source

https://www.icao.int/publications/DOC8643/Pages/Search.aspx