Export FlightRadar24 position reports from flights flying around an airport

Extract FlightRadar24 positions within a distance from the aerodrome for qualified airport movements (arrivals or departures or all)

Usage

export_positions_at_airport_fr24(
  wef,
  til,
  apt,
  lon_apt,
  lat_apt,
  flow = "ALL",
  radius = 40
)

Arguments

wef

(UTC) timestamp of With Effect From (included)

til

(UTC) timestamp of TILl instant (excluded)

apt

IATA airport code, i.e. PSA for Pisa "Galileo Galilei"

lon_apt

airport longitude (decimal degrees, WGS84)

lat_apt

airport latitude (decimal degrees, WGS84)

flow

the flow of flights: "ARR" for arrivals, "DEP" for departures, "ALL" for both [default "ALL"]

radius

radius around airport to keep position reports (nautical miles [NM])

Value

dataframe of ADS-B position reports

Details

You need to store your credentials to access the FR24 tables in the following environment variables:

PRU_FR24_USR for the user id

PRU_FR24_PWD for the password

PRU_FR24_DBNAME for the database name

See also

Examples

if (FALSE) { # \dontrun{
# half a day (UTC times, not local ones!) worth of all movements at Stavanger Airport,
# Sola, Sweden on 1st Sep 2017
export_positions_at_airport_fr24("2017-09-01T00:00:00",
                                 "2017-09-01T12:00:00",
                                 "SVG",
                                 5.638, 58.877)

# all arrivals within 50 NM on 25th Sep 2018 at Pisa Airport, Pisa, Italy
export_positions_at_airport_fr24("2018-09-25",
                                 "2018-09-26",
                                 "PSA", 10.39270, 43.68390,
                                 flow = "ARR",
                                 radius = 50)
} # }