Extract FR24 flights list for a time interval at an airport

You need to store your credentials to access the FR24 tables in the following environment variables:

PRU_FR24_USR for the user id

PRU_FR24_PWD for the password

PRU_FR24_DBNAME for the database name

Usage

export_positions_fr24(wef, til)

Arguments

wef

(UT) timestamp of With Effect From (included)

til

(UT) timestamp of TILl instant (excluded)

Value

a dataframe of position reports

See also

Examples

if (FALSE) { # \dontrun{
export_positions_fr24("2017-09-01T00:00:00Z", "2017-09-01T06:00:00Z")
export_positions_fr24("2017-09-01 10:05:00Z", "2017-09-01T11:33:00")
} # }