Extract level segment from a list of position reports.

It is assumed that the dataset is grouped by flight id and that position reports are sorted by ascending timestamp.

Usage

extract_segment(prs, vert_limit = 5)

Arguments

prs

dataset containing position reports. Two variables are used: timestamp and altitude (feet)

vert_limit

the maximum absolute vertical rate for a segment to be considered level (default = 5 feet/s)

Value

a dataset of level segments. A level segment is defined by its beginning and end 4D position.

See also

Examples

if (FALSE) {
cumulative_time(poss)
f <- cprs %>%
 filter(callsign == "RYR40WJ") %>%
 arrange(timestamp_track) %>%
 rename(timestamp = timestamp_track) %>%
 mutate(altitude = flight_level * 100)
 extract_segment(f)
}