Parse latitude in ICAO format

The format is as described in ICAO Doc 8126 section 5.5.2 bullet d):

  In the indication of the geographical coordinates of
  a location:
  - the latitude should be given first;
  - symbols for degrees, minutes or seconds should
    be omitted;
  - two digits should always be used in expressing
    values of less than 10 degrees of latitude; and
  - three digits should always be used in expressing
    values of less than 100 degrees of longitude.
  For example, 050735N 0652542W means five
  degrees, seven minutes and thirty-five seconds
  North, sixty-five degrees, twenty-five minutes and
  forty-two seconds West.

What is typical in AIP pubblications is the additional part of decimal seconds, like 554718.23N.

Usage

parse_lat_icao(lat)

Arguments

lat

a string representation for latitude, i.e. 554718N or 554718.23N

Value

a latitude in decimal degrees

Details

This format is used in AIP pubblications, for examples for runway threshold coordinates, stand positions, etc. This is used also in ALL_FT+ files from NM.

See also

Examples

parse_lat_icao("554718N")
#> [1] 55.78833