Sample FlightRadar24's archived feed ADS-B position reports data.

A dataset containing FlightRadar24's archived ADS-B position reports (combined with flight information of flts) for 41 flights. There are 20 flights on 5th and 21 on 6th Feb 2017.

Usage

poss

Format

A data frame with 9290 rows and 19 variables. The variables are a joint of flight information from flts with the following ones:

flight_id

unique identifier for the flight leg (integer)

timestamp

datetime for the position report (convertion of snapshot_id)

snapshot_id

time of sample as number of seconds that have elapsed since 00:00:00 Coordinated Universal Time (UTC), Thursday, 1 January 1970

altitude

altitude in feet (integer), as supplied by the Mode-S or ADS-B transponder

heading

degrees from North, 0-359, (integer), data supplied by ADS-B but calculated for mode-S transponders and tracked using multi-lateration (MLAT)

latitude

latitude in decimal degrees, data supplied by ADS-B but calculated for mode-S transponders and tracked using multi-lateration (MLAT)

longitude

longitude in decimal degrees, data supplied by ADS-B but calculated for mode-S transponders and tracked using multi-lateration (MLAT)

radar_id

unique identifier of primary ground receiver

speed

ground speed in knots (integer), data supplied by ADS-B but calculated for mode-S transponders and tracked using multi-lateration (MLAT)

squawk

code broadcast by airplane represented as four octal digits, data supplied directly from the mode-S or ADS-B transponder

aircraft_id

mode-S address in hexadecimal, as supplied by the Mode-S or ADS-B transponder

reg

unique aircraft identifier (text)

equip

aircraft type (e.g. B733, as per ICAO classification?), GLID for gliders, GROUND for ground vehicles

callsign

e.g. CSA190, as supplied by the Mode-S or ADS-B transponder

flight

flight number, e.g. OK190

schd_from

IATA code for scheduled departure airport

schd_to

IATA code for scheduled arrival airport

real_to

IATA code for actual arrival airport (when diverted)

date

date of the flight (what about the ones crossing midnight?)

Source

FlightRadar24