Read FR24's archived flights information for a day.

read_flights_fr24 returns all flights as archived by FR24 for a given day (see path parameter for how the date is encoded).

read_flights_fr24(path)

Arguments

path
the filename of the <yyyymmdd>_flights<eventually>.csv file. The <yyyymmdd> part of the path's basename is extracted into the date field.

Value

a data frame of the archived flights that have flown on <yyyymmdd>. Each flight has the following attributes (all optional, i.e. eventually empty, but flight_id):

date date of the flight

flight_id unique decimal identifier for the flight leg

aircraft_id mode-S address in hexadecimal

reg unique aircraft text identifier

equip equipment, e.g. B733

callsign callsign, e.g. NAX11S

flight flight number, e.g. DY311

schd_from IATA code for scheduled departure airport, e.g. KKN

schd_to IATA code for scheduled arrival airport

real_to IATA code for actual arrival airport (when diverted)

Details

The flights listed are the ones known to FR24 and fitting the the bounding box as contractually agreed.

See also

Examples

# NOT RUN {
# flights, as recorded by FR24, on 20170205 (5th Feb 2017)
flt_fr24_file <- system.file("extdata",
                         "20170205_flights.csv"
                         package = "trrrj")
read_flights_fr24(flt_fr24_file)
# }