Tidy eval helpers

enquo() delays the execution of one or several function arguments. enquo() returns a single quoted expression, which is like a blueprint for the delayed computation.

expr() quotes a new expression locally. It is mostly useful to build new expressions around arguments captured with enquo() or enquos(): expr(mean(!!enquo(arg), na.rm = TRUE)).

as_name() transforms a quoted variable name into a string. Supplying something else than a quoted variable name is an error.

as_label() transforms any kind of R object, including quoted function calls and vectors, into a string. Its purpose is to summarise that object into a single label. That label is often suitable as a default name.

If you don't know what a quoted expression contains (for instance expressions captured with enquo() could be a variable name, a call to a function, or an unquoted constant), then use as_label().

To learn more about tidy eval and how to use these tools, visit https://tidyeval.tidyverse.org and the Metaprogramming section of Advanced R.