Sample Correlated Position Report (CPR) data.

A dataset containing CPR's for 41 flights flying over Europe. One on 4th , 19 on 5th and 21 on 6th Feb 2017.

A dataset containing CPR's for 41 flights flying over Europe. One on 4th , 19 on 5th and 21 on 6th Feb 2017.

cprs

cprs

Format

A data frame with 53940 rows and 10 variables:

cpm_id

CPR Message (CPM) line number

tact_id

TACT Id, TACT (a.k.a. ETFMS) is an NM system

timestamp_etfms

time of CPM reception by the ETFMS system

timestamp_track

time of track

block

block number. ETFMS internal use

record

record number. ETFMS internal use

entry_node_sac

Entry Node (EN) system area code (SAC). To avoid ambiguity in the exchange of Surveillance related data, each system using the ASTERIX data format gets assigned a unique identifier composed of two values called `SAC/SIC'. See https://www.eurocontrol.int/services/system-area-code-list

entry_node_sic

Entry Node (EN) system identifier code (SIC). To avoid ambiguity in the exchange of Surveillance related data, each system using the ASTERIX data format gets assigned a unique identifier composed of two values called `SAC/SIC'. The System Identification Code (SIC) is allocated nationally by the responsible Air Traffic Services Organisation. It identifies each individual system (surveillance sensor, surveillance data processing system, etc) within the respective area defined by the SAC. See https://www.eurocontrol.int/services/system-area-code-list

callsign

Callsign for the flight as provided in the FPL (Flight PLan). It may be the registration marking of the aircraft or the ICAO designator for the aircraft operating agency followed by the flight identification

adep_icao

ICAO location identifier of the Airport of Departure (ADEP)

ades_icao

ICAO location identifier of the Airport of Destination (ADES)

eobt

Estimate Take-Off Date and Time (EOBT), the estimated time at which the aircraft will commence movement associated with departure

longitude

longitude (decimal degrees)

latitude

latitude (decimal degrees)

flight_level

flight level of the aircraft. Flight levels are surfaces of constant atmospheric pressure which are related to a specific pressure datum, 1013.2 HP (Hecto-Pascal). The expression `Flight level times 100' is sometimes, not quite correctly, referred to as altitude in feet

track_service

Determines whether the CPR is the first (Begin), an intermediate (Continuing) or the last (End) CPR sent by the corresponding system for the relevant flight. (Begin_And_End is also possible)

ssr_code

A 4-Digit octal code used in the transponder to identify an aircraft (SSR Mode 3/A). See https://en.wikipedia.org/wiki/Aviation_transponder_interrogation_modes

track_speed

calculated ground velocity (knots) based on the previous radar position

track_heading

calculated heading of aircraft with respect to the magnetic North (decimal degrees)

climb_rate

Climb (positive) or descend (negative) rate (knots)

track_vertical_mode

a categorical value for the rate of climb, it can be one of CLIMB, DESCENT, LEVEL_FLIGHT or UNDETERMINED

ifps_id

a unique flight plan identifier assigned by the IFPS system

aircraft_address

a unique identification of the aircrafts frame (24-bit ICAO aircraft address)

Source

EUROCONTROL's Network Manager

EUROCONTROL's Network Manager