Read FR24's archived positions information for a flight.

read_positions_fr24 returns all position reports as archived by FR24 for a given flight_id flown on a specific day (see path parameter for how the date and flight ID are encoded).

read_positions_fr24(path)

Arguments

path
the filename of the <yyyymmdd>_<flight_id>_positions.csv file

Value

a data frame of the archived flight's positions for date <yyyymmdd>. Each position report has the following attributes (all mandatory, i.e. non empty, but speed, squawk and vert_speed):

flight_id FR24 unique (within a day?) flight leg identifier

timestamp snapshot_id converted to a date/time

snapshot_id time of sample in Unix epoch seconds, i.e. since 1st Jan 1970 00:00:00 UTC

altitude altitude (feet). Barometric relative to ISO 1013 hPa pressure in flight, always 0 on ground.

heading 0-359 degrees

latitude latitude in decimal degrees

longitude longitude in decimal degrees

radar_id unique identifier of primary ground receiver

speed ground speed (knots)

squawk code broadcast by airplane, represented as four octal digits

vert_speed vertical speed (feet/min). Calculated from rate of altitude changes, 0 for level flight or empty if unavailable

on_ground 1 for on ground, 0 for airborne. Calculated from low altitude and ground speed

Details

The positions listed are the ones known to FR24 and fitting the the bounding box as contractually agreed.

See also

Examples

# NOT RUN {
# positions, as recorded by FR24, for flight 207535493 on 20170205 (5th Feb 2017)
pos_fr24_file <- system.file("extdata",
                             "20170205_positions",
                             "20170205_207535493.csv"
                             package = "trrrj")
read_positions_fr24(pos_fr24_file)
# }