Read SO6 file

The SO6 file has variables of two types: * Segment-specific data (latitude/longitude, time/data information, ...) * Flight meta data (flight id, callsign, aircraft type, ...)

They are marked S or F respectively in the table below.

Name	Description	Type	Kind

segment_id	segment Id	char	S

adep	Departing aerodrome (ICAO) ID	char	F

ades	Destination aerodrome (ICAO) ID	char	S

aircraft_type	ICAO aircraft type	char	S

segment_hhmm_begin	Segment's begin timestamp (hhmmss)	char	S

segment_hhmm_end	Segment's end timestamp (hhmmss)	char	S

segment_fl_begin	Segment's begin flight level	int	S

segment_fl_end	Segment's end flight level	int	S

status	Segment's status (0=climb, 1=descent, 2=cruise)	factor	S

callsign	Flight call sign	char	F

segment_date_begin	Segment's begin date (YYMMDD)	char	S

segment_date_end	Segment's end date (YYMMDD)	char	S

segment_latitude_begin	Segment's begin latitude (Min decimal)	char	S

segment_longitude_begin	Segment's begin longitude (Min decimal)	char	S

segment_latitude_end	Segment's end latitude (Min decimal)	char	S

segment_longitude_end	Segment's end longitude (Min decimal)	char	S

flight_id	Flight ID	int	F

sequence	Segment's sequence	int	S

segment_length	Segment's length	double	S

segment_parity	Segment's parity	int	S

segment_timestamp_begin	Segment's begin timestamp	datetime	S

segment_timestamp_end	Segment's end timestamp	datetime	S

point_id_begin	Segment's begin point ID	char	S

point_id_end	Segment's end point ID	char	S

read_so6(filename, delim = " ")

Arguments

filename
the file containing SO6 trajectories

delim
the field delimiter (default: " " [blank])

Value

a data frame where date and time are combined in a single datetime and longitude and latitude are in decimal degrees.

See also

Examples

if (FALSE) {
so6file <- system.file("extdata",
                       "TRAFFIC_20180630_reduced.so6",
                       package = "trrrj")
read_so6(so6file)
}